Add wheel building.
