+++ MANIFEST.in
@@ -1,0 +1,2 @@
+include documentation/*
+include mini_civ/mini_civ/res/*
