+++ mini_civ/mini_civ/game.py
@@ -5,10 +5,11 @@ sys.path.append(os.path.normpath(os.path.join
 
 from src.game import Game
 
-WIDTH = 800
-HEIGHT = 650
-FPS = 30
+def start_game():
+    WIDTH = 800
+    HEIGHT = 650
+    FPS = 30
 
-play = Game(WIDTH, HEIGHT, FPS)
-play.mainloop()
-play.quit()
+    play = Game(WIDTH, HEIGHT, FPS)
+    play.mainloop()
+    play.quit()
