+++ mini_civ/mini_civ/res/base-hex.png
