+++ mini_civ/mini_civ/res/cursor1_rs2.png
