+++ mini_civ/mini_civ/res/desert-hex.png
