+++ mini_civ/mini_civ/res/frame_button1.png
