+++ mini_civ/mini_civ/res/frame_global5.png
