+++ mini_civ/mini_civ/res/ice-hex.png
