+++ mini_civ/mini_civ/res/plains-hex.png
