+++ mini_civ/mini_civ/res/po/game.pot
