+++ mini_civ/mini_civ/res/po/ru/LC_MESSAGES/game.po
