+++ mini_civ/mini_civ/res/spearman.png
