+++ mini_civ/mini_civ/res/spearman_attacked.png
