+++ mini_civ/mini_civ/res/spearman_red.png
