+++ mini_civ/mini_civ/res/townhall.png
