+++ mini_civ/mini_civ/res/townhall_attacked.png
