+++ mini_civ/mini_civ/res/townhall_blue.png
