+++ mini_civ/mini_civ/res/townhall_red.png
