+++ mini_civ/src/cells/plain_cell.py
