+++ mini_civ/src/units/creatures/spearman.py
