+++ mini_civ/src/visual/vis_cell.py
