+++ mini_civ/src/visual/vis_map.py
@@ -275,12 +275,12 @@ def generate_map(game_state: Game_state,
     gamemap.gen_water(banned_cells)
 
     townhall_red_img = pygame.image.load(os.path.join(game_folder,
-                                         'res/townhall_red.png')).convert()
+                                         'mini_civ', 'res', 'townhall_red.png')).convert()
     townhall_blue_img = pygame.image.load(os.path.join(game_folder,
-                                          'res/townhall_blue.png')).convert()
+                                          'mini_civ', 'res', 'townhall_blue.png')).convert()
     townhall_attacked = pygame.image.load(
         os.path.join(game_folder,
-                     'res/townhall_attacked.png')).convert()
+                     'mini_civ', 'res', 'townhall_attacked.png')).convert()
     for building in red_stat.get_buildings():
         building.add_vis_unit(townhall_red_img)
         building.get_vis_unit().set_attacked(townhall_attacked)
@@ -297,12 +297,12 @@ def generate_map(game_state: Game_state,
         building.vis_unit.set_immovable(True)
 
     spearman_red_img = pygame.image.load(os.path.join(game_folder,
-                                         'res/spearman_red.png')).convert()
+                                         'mini_civ', 'res', 'spearman_red.png')).convert()
     spearman_blue_img = pygame.image.load(os.path.join(game_folder,
-                                          'res/spearman_blue.png')).convert()
+                                          'mini_civ', 'res', 'spearman_blue.png')).convert()
     spearman_attacked = pygame.image.load(
         os.path.join(game_folder,
-                     'res/spearman_attacked.png')).convert()
+                     'mini_civ', 'res', 'spearman_attacked.png')).convert()
     for unit in red_stat.get_units():
         unit.add_vis_unit(spearman_red_img)
         unit.get_vis_unit().set_attacked(spearman_attacked)
