+++ tests/deploy_test.py
@@ -1,0 +1,2 @@
+from mini_civ.game import start_game
+start_game()
